requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.1
dj-database-url==0.5.0
Django==4.1
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.1.0
psycopg2
djangorestframework
ipdb
djangorestframework-simplejwt